Restrict access to boards (#676)
* Refactor user board and user board column views

* Simplify auth checks for board access

* Refactor view access control in user_board views

* front-end tests for controlled board access

* Remove unnecessary is_staff=True instances and add explanatory comment

* Update login page and fix test cases

* Refactor user board access control and views

This commit refactors the user board access control and views in the frontend. The `TestUserBoardControlledAccess` class has been renamed to `TestUserBoardAuthorization` for clarity. The `user_passes_test_or_forbidden` decorator has been added to the views to check if the user has permission to access the page and return a 403 Forbidden response if not. This ensures that unauthorized users cannot view other users' boards.